+++ arrus/core/devices/us4r/hv/HighVoltageSupplier.h
@@ -24,13 +24,12 @@ public:
 
     HighVoltageSupplier(const DeviceId &id, HVModelId modelId);
 
-    void setVoltage(const std::vector<HVVoltage> &voltages) {
-        std::vector<IHVVoltage> us4RVoltages;
-        std::transform(std::begin(voltages), std::end(voltages), std::back_insert_iterator(us4RVoltages),
-                       [](const HVVoltage &v) { return IHVVoltage(v.getVoltageMinus(), v.getVoltagePlus()); });
+    void setVoltage(const std::vector<IHVVoltage> &voltages) {
+        ARRUS_REQUIRES_TRUE(voltages.size() == 2 || voltages.size() == 1,
+                            "The vector of voltages should contain exactly two values!");
         try {
             getIHV()->EnableHV();
-            getIHV()->SetHVVoltage(us4RVoltages);
+            getIHV()->SetHVVoltage(voltages);
         } catch (const ::us4us::ValidationException &) {
             // Disable HV and Propage validation errors.
             try {
@@ -53,14 +52,26 @@ public:
                                         "message: '{}', trying once more.",
                                         e.what()));
             getIHV()->EnableHV();
-            getIHV()->SetHVVoltage(us4RVoltages);
+            getIHV()->SetHVVoltage(voltages);
         }
     }
 
+    /**
+     * Returns the default measured voltage.
+     * For the OEM HVPS, this is the voltage measured on the rail 0 / amplitude 2.
+     */
     unsigned char getVoltage() { return getIHV()->GetHVVoltage(); }
 
+    /**
+     * Returns the default measured voltage.
+     * For the OEM HVPS, this is the voltage measured on the rail 0 / amplitude 2.
+     */
     float getMeasuredPVoltage() { return getIHV()->GetMeasuredHVPVoltage(); }
 
+    /**
+     * Returns the default measured voltage.
+     * For the OEM HVPS, this is the voltage measured on the rail 0 / amplitude 2.
+     */
     float getMeasuredMVoltage() { return getIHV()->GetMeasuredHVMVoltage(); }
 
     void disable() {